Separate email into its own service, fix failing test
