+++ server/src/main/java/com/objectcomputing/checkins/notifications/email/EmailSender.java
@@ -1,11 +1,7 @@
 package com.objectcomputing.checkins.notifications.email;
 
-import java.util.List;
-
 public interface EmailSender {
     void sendEmail(String subject, String content, String... recipients);
 
     boolean sendEmailReceivesStatus(String subject, String content, String... recipients);
-
-    List<Email> sendAndSaveEmail(String subject, String content, String... recipients);
 }
