+++ server/src/main/java/com/objectcomputing/checkins/notifications/email/MailJetSender.java
@@ -4,13 +4,8 @@ import com.mailjet.client.MailjetClient;
 import com.mailjet.client.MailjetRequest;
 import com.mailjet.client.MailjetResponse;
 import com.mailjet.client.errors.MailjetException;
-import com.mailjet.client.errors.MailjetServerException;
 import com.mailjet.client.errors.MailjetSocketTimeoutException;
 import com.mailjet.client.resource.Emailv31;
-import com.objectcomputing.checkins.exceptions.PermissionException;
-import com.objectcomputing.checkins.services.memberprofile.MemberProfile;
-import com.objectcomputing.checkins.services.memberprofile.MemberProfileRepository;
-import com.objectcomputing.checkins.services.memberprofile.currentuser.CurrentUserServices;
 import io.micronaut.context.annotation.Property;
 import io.micronaut.context.annotation.Requires;
 import org.json.JSONArray;
@@ -18,8 +13,6 @@ import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import javax.inject.Singleton;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -29,10 +22,7 @@ import java.util.stream.Collectors;
 public class MailJetSender implements EmailSender {
 
     private static final Logger LOG = LoggerFactory.getLogger(MailJetSender.class);
-    private MailjetClient client;
-    private final EmailRepository emailRepository;
-    private final CurrentUserServices currentUserServices;
-    private final MemberProfileRepository memberProfileRepository;
+    private final MailjetClient client;
 
     public static final String FROM_ADDRESS = "mail-jet.from_address";
     public static final String FROM_NAME = "mail-jet.from_name";
@@ -42,23 +32,19 @@ public class MailJetSender implements EmailSender {
     private final String fromName;
 
     public MailJetSender(MailjetClient client,
-                         EmailRepository emailRepository,
-                         CurrentUserServices currentUserServices,
-                         MemberProfileRepository memberProfileRepository,
                          @Property(name = FROM_ADDRESS) String fromAddress,
                          @Property(name = FROM_NAME) String fromName) {
         this.client = client;
-        this.emailRepository = emailRepository;
-        this.currentUserServices = currentUserServices;
-        this.memberProfileRepository = memberProfileRepository;
         this.fromAddress = fromAddress;
         this.fromName = fromName;
     }
 
-    public void setEmailClient(MailjetClient client) {
-        this.client = client;
-    }
-
+    /**
+     * Helper method to divide a list of recipients into smaller lists (batches) of recipients
+     * Enables sending emails to more recipients than permitted by a single MailJet API call
+     * @param recipients List of recipient email addresses
+     * @return list of {@link JSONArray}
+     */
     public static List<JSONArray> getEmailBatches(String... recipients) {
         List<String> recipientList = new ArrayList<>(List.of(recipients));
         Collections.sort(recipientList);
@@ -87,6 +73,7 @@ public class MailJetSender implements EmailSender {
      * This call sends a message to the given recipient with attachment.
      * @param subject, {@link String} Subject of email
      * @param content {@link String} Contents of email
+     * @param recipients List of recipient email addresses
      */
     @Override
     public void sendEmail(String subject, String content, String... recipients) {
@@ -136,42 +123,5 @@ public class MailJetSender implements EmailSender {
             return false;
         }
         return true;
-    }
-
-    @Override
-    public List<Email> sendAndSaveEmail(String subject, String content, String... recipients) {
-
-        List<Email> sentEmails = new ArrayList<>();
-
-        if (!currentUserServices.isAdmin()) {
-            throw new PermissionException("You are not authorized to do this operation");
-        }
-
-        LocalDateTime sendDate = LocalDateTime.now();
-        boolean status = sendEmailReceivesStatus(subject, content, recipients);
-
-        UUID senderId = currentUserServices.getCurrentUser().getId();
-
-        if (status) {
-            for (String recipientEmail : recipients) {
-                Optional<MemberProfile> recipient = memberProfileRepository.findByWorkEmail(recipientEmail);
-                if (recipient.isPresent()) {
-                    // Only send emails to unterminated members
-                    LocalDate terminationDate = recipient.get().getTerminationDate();
-                    if (terminationDate == null || terminationDate.isAfter(LocalDate.now())) {
-                        UUID recipientId = recipient.get().getId();
-                        Email email = new Email(subject, content, senderId, recipientId, sendDate, LocalDateTime.now());
-                        Email savedEmail = emailRepository.save(email);
-                        sentEmails.add(savedEmail);
-                    } else {
-                        LOG.warn(String.format("Prevented sending email to terminated member: %s", recipientEmail));
-                    }
-                }
-            }
-        } else {
-            throw new RuntimeException("Failed to send emails");
-        }
-
-        return sentEmails;
     }
 }
