+++ server/src/main/java/com/objectcomputing/checkins/services/email/EmailServices.java
@@ -1,0 +1,9 @@
+package com.objectcomputing.checkins.services.email;
+
+import java.util.List;
+
+public interface EmailServices {
+
+    List<Email> sendAndSaveEmail(String subject, String content, String... recipients);
+
+}
