+++ server/src/test/java/com/objectcomputing/checkins/services/fixture/RepositoryFixture.java
@@ -1,6 +1,6 @@
 package com.objectcomputing.checkins.services.fixture;
 
-import com.objectcomputing.checkins.notifications.email.EmailRepository;
+import com.objectcomputing.checkins.services.email.EmailRepository;
 import com.objectcomputing.checkins.services.action_item.ActionItemRepository;
 import com.objectcomputing.checkins.services.agenda_item.AgendaItemRepository;
 import com.objectcomputing.checkins.services.checkin_notes.CheckinNoteRepository;
